jsonapi/requirements.txt
Django==3.1.5
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
whitenoise==5.2.0
dj_database_url==0.5.0
djongo==1.3.6